Adds an API fetching
